% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eexbqfr1_signal.R
\name{debqfr1_model}
\alias{debqfr1_model}
\title{Trade signal for German Power Futures. Front Base Quarter contract}
\usage{
debqfr1_model(eexb, api2q)
}
\arguments{
\item{eexb}{an xts object. this is the front base quarter contract traded on EEX.
https://www.eex.com/en/market-data/power/futures/phelix-deat-futures#!/2017/09/08}

\item{api2q}{an xts object. This is the front quarter contract traded/
cleared on ICE.
https://www.theice.com/products/243/API2-Rotterdam-Coal-Futures/data}
}
\value{
a logical object. (1 or -1) the trade signal.
}
\description{
Trade signal for German Power Futures. Front Base Quarter contract
}
